Draw a quad and wireframe mode
Reverted the color and position vertex buffers code. Use a single vertex buffer and draw two traingles to make it a Quad, changed the glDrawArrays() to take 6 vertices.

Implemented a wireframe mode on click of W key on the keyboard, which draws outline and on toggle of W key it changes to fill mode.